+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 ### Improvements
 - Send "Monthly Phone/SMS/WA Limit Reached" notice to all team members (not just owner)
 - Include notification text in the "Monthly Phone/SMS/WA Limit Reached" notice
+- Add a note about unset ping key in the check details page (#1295)
 
 ### Bug Fixes
 - Fix WebhookValidator to handle input that contains square brackets